Added another scene, 'Prototype Room'
Added a new scene and added assets that were created in Blender. All assets from Blender are in the Blender Assets folder. The commit prior to this one has the same name and similar description.. Ignore it, I made a mistake :( I pushed the game but with not much changed to it...