Added mirrors and doors. Adjusted battery pick up
Doors and mirrors are now present in the game. Doors are opened and closed by clicking on them with the left mouse button. You now pick up batteries with the left mouse button. Now we can place batteries on desks, beds, couches, etc. around the map and have the player collect them.